+++ tests/utils/utils_workers.py
@@ -1,4 +1,4 @@
-from tests.utils import MockOrchestrator
+from tests.utils.utils import MockOrchestrator
 
 
 class MCWorkerOrchestrator(MockOrchestrator):
